requirements.txt
Django==3.1.7
mysqlclient==2.0.1
django-mysql==3.8.1
django-redis==4.12.1
celery==5.0.5

pytest==6.2.2
pytest-django==4.1.0
factory-boy==3.2.0